+++ src/components/CurriculumComponents/CurricTimetablingUnits/index.tsx
@@ -3,38 +3,82 @@ import {
   OakBox,
   OakFlex,
   OakHeading,
+  OakInformativeModal,
   OakMaxWidth,
+  OakSecondaryButton,
 } from "@oaknational/oak-components";
 import { ThemeProvider } from "styled-components";
-import { useMemo } from "react";
+import { useMemo, useState } from "react";
 
 import { CurricTimetableHeader } from "../CurricTimetableHeader";
-import { CurricShowSteps } from "../CurricShowSteps";
+import CurricTimetablingFilters from "../CurricTimetablingFilters";
 
 import { useTimetableParams } from "@/utils/curriculum/timetabling";
-import { parseSubjectPhaseSlug } from "@/utils/curriculum/slugs";
-import { Unit } from "@/utils/curriculum/types";
+import { CurriculumFilters, Unit } from "@/utils/curriculum/types";
 import oakTheme from "@/styles/theme";
 import {
   fetchSubjectPhasePickerData,
   formatCurriculumUnitsData,
 } from "@/pages-helpers/curriculum/docx/tab-helpers";
+import { CurriculumSelectionSlugs } from "@/utils/curriculum/slugs";
+
+function getDefaultName(
+  data: ReturnType<typeof useTimetableParams>[0],
+  units: Unit[],
+) {
+  if (data.name && data.name !== "") {
+    return data.name;
+  }
+  return `Your ${units[0]?.subject ?? ""} timetable`;
+}
 
 type CurricTimetablingUnitsProps = {
-  subjectPhaseSlug: string;
+  slugs: CurriculumSelectionSlugs;
   units: Unit[];
-  curriculumPhaseOptions: Awaited<
-    ReturnType<typeof fetchSubjectPhasePickerData>
-  >;
+  curriculumPhaseOptions:
+    | ReturnType<typeof fetchSubjectPhasePickerData>
+    | Awaited<ReturnType<typeof fetchSubjectPhasePickerData>>;
 };
 export const CurricTimetablingUnits = ({
-  subjectPhaseSlug,
   units,
+  slugs,
 }: CurricTimetablingUnitsProps) => {
+  const [modalOpen, setModalOpen] = useState(false);
   const [data] = useTimetableParams();
-  const slugs = parseSubjectPhaseSlug(subjectPhaseSlug)!;
+  const [filters, setFilters] = useState<CurriculumFilters>(() => {
+    return {
+      years: [data.year ?? "1"],
+      childSubjects: [],
+      subjectCategories: [],
+      tiers: [],
+      threads: [],
+      pathways: [],
+    };
+  });
 
-  const unitData = useMemo(() => formatCurriculumUnitsData({ units }), [units]);
+  const unitDataPre = useMemo(
+    () => formatCurriculumUnitsData({ units }),
+    [units],
+  );
+
+  const unitData = useMemo(() => {
+    return {
+      ...unitDataPre,
+      yearOptions:
+        data.year && unitDataPre.yearOptions.includes(data.year)
+          ? [data.year]
+          : [],
+    };
+  }, [data, unitDataPre]);
+
+  const onEditDetails = () => {
+    setModalOpen(true);
+  };
+
+  const onCopyLink = () => {
+    const urlToCopy = window.location.href;
+    navigator.clipboard.writeText(urlToCopy);
+  };
 
   return (
     <>
@@ -42,20 +86,48 @@ export const CurricTimetablingUnits = ({
       <ThemeProvider theme={oakTheme}>
         <OakFlex $flexDirection={"column"} $pa={"inner-padding-xl5"}>
           <CurricTimetableHeader
-            titleSlot={`Year ${data.year} ${slugs.subjectSlug}`}
+            titleSlot={getDefaultName(data, units)}
             illustrationSlug={"magic-carpet"}
             additionalSlot={
-              <OakBox $maxWidth={"all-spacing-20"}>
-                <CurricShowSteps numberOfSteps={2} currentStepIndex={1} />
-              </OakBox>
+              <OakFlex $maxWidth={"all-spacing-20"} $gap={"all-spacing-4"}>
+                <OakSecondaryButton
+                  iconName="copy"
+                  isTrailingIcon={true}
+                  onClick={onEditDetails}
+                >
+                  Edit details
+                </OakSecondaryButton>
+                <OakSecondaryButton
+                  iconName="edit"
+                  isTrailingIcon={true}
+                  onClick={onCopyLink}
+                >
+                  Copy link
+                </OakSecondaryButton>
+              </OakFlex>
             }
           />
         </OakFlex>
 
+        <OakInformativeModal
+          isOpen={modalOpen}
+          onClose={() => setModalOpen(false)}
+          closeOnBackgroundClick={true}
+        >
+          <OakBox data-testid="edit-details-modal">Edit details modal</OakBox>
+        </OakInformativeModal>
+
         <OakMaxWidth $ph={"inner-padding-xl5"}>
           <OakFlex $flexDirection={"row"}>
             <OakFlex $width={"all-spacing-21"} $flexDirection={"column"}>
-              <OakHeading tag="h2">Filter and highlight</OakHeading>
+              <CurricTimetablingFilters
+                filters={filters}
+                onChangeFilters={setFilters}
+                slugs={slugs}
+                data={unitData}
+              />
+
+              <OakHeading tag="h2">Debug</OakHeading>
               <pre>{JSON.stringify(data, null, 2)}</pre>
             </OakFlex>
             <OakFlex $flexGrow={1}>