+++ src/context/Search/useSuggestedFilters.tsx
@@ -31,7 +31,11 @@ export function useSuggestedFilters({
   enabled: boolean;
 }): SuggestedFilters {
   const params = useMemo(
-    () => new URLSearchParams({ searchTerm: normalizeTerm(term) }).toString(),
+    () =>
+      new URLSearchParams({
+        v: "1",
+        searchTerm: normalizeTerm(term),
+      }).toString(),
     [term],
   );
 
@@ -70,6 +74,7 @@ export function useSuggestedFilters({
     return {
       searchFilters: convertSearchIntentToFilters(data),
       status: "success",
+      data,
     };
   }